Restructures main func for referencing assets
The MEL procedure has been simplified. Initially,
it was trying to reference the selected file by
calling the defined Python functions in the module.
However, MEL doesn't seem to recognize the functions
because they're defined within the script, not in
__main__ as it probably expects. Now, it simply
assigns the file path to a global variable
"$_passToPython", then maya.mel.eval simply grabs
its value. The referencing is done in Python